requirements.txt
asgiref==3.4.1
boto3==1.18.31
botocore==1.21.31
dj-database-url==0.5.0
Django==3.2.6
django-cors-headers==3.8.0
django-heroku==0.3.1
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
jmespath==0.10.0
Pillow==8.3.1
psycopg2==2.9.1
PyJWT==2.0.1
python-dateutil==2.8.2
pytz==2021.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.3.0
